server/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Cors
Flask-Mail
Flask-Swagger-UI
Flask-Uploads
bcrypt
celery
mysqlclient